+++ test/js/remote/vrpcRemoteTest.js
@@ -4,7 +4,7 @@
 
 const { assert } = require('chai')
 const EventEmitter = require('events')
-const VrpcRemote = require('../../vrpc/VrpcRemote')
+const { VrpcRemote } = require('../../../index')
 
 const emitter = new EventEmitter()
 const newEntries = []
@@ -26,7 +26,7 @@ describe('An instance of the VrpcRemote class', () => {
   })
   it('should return available classes and functions', async () => {
     const availabilities = await vrpc.getAvailabilities()
-    console.log('Availabilities:', JSON.stringify(availabilities, null, 2))
+    // console.log('Availabilities:', JSON.stringify(availabilities, null, 2))
     // console.log('Domains:', await vrpc.getAvailableDomains())
     // console.log('Agents:', await vrpc.getAvailableAgents())
     // console.log('Classes:', await vrpc.getAvailableClasses('js'))